requirements.txt
requests
beautifulsoup4
pandas
pyyaml
tqdm
lxml
psutil
streamlit
